requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
geoalchemy2
passlib[bcrypt]
python-jose
python-dotenv